fix a bunch of specs... also worked on sources I think

+++ app/graphql/mutations/unfavorite_instance.rb
@@ -14,9 +14,7 @@ module Mutations
 
     def resolve(klass:, id:, user_id:, favorite_id:)
       favorite = Favorite.find(favorite_id)
-      if favorite.user_id == user_id && favorite.klass == klass
-        favorite.destroy
-      end
+      favorite.destroy if favorite.user_id == user_id && favorite.klass == klass
       if favorite.persisted?
         {
           errors: ['unable to destroy instance'],
